🧹 [code health] Refactor app/create-session.tsx and update env docs
Extracts multiple internal components from `app/create-session.tsx` to separate files within `components/jules/create-session/` to improve code health and reduce file size.
Updates `constants/api-key-context.tsx` to utilize `process.env.JULES_API_KEY` for local development.
Updates README files and DEVELOPMENT documentation to clarify how to use the `.env` variable for the Jules API Key.
